[TM-3677] Refactor project and project report entities by removing unused fields, updating form handling, and enhancing tests for project report persistence.
