+++ apps/entity-service/src/entities/form-data.service.ts
@@ -8,7 +8,6 @@ import {
   FundingProgramme,
   I18nTranslation,
   Media,
-  ProjectReport,
   Stage,
   UpdateRequest
 } from "@terramatch-microservices/database/entities";
@@ -17,7 +16,6 @@ import {
   EntityModel,
   EntityType,
   FormModel,
-  FormModelType,
   formModelType,
   getOrganisationId,
   getProjectId,
@@ -68,9 +66,7 @@ export class FormDataService {
       model.updateRequestStatus = newUpdateRequest.status;
       await model.save();
     } else {
-      const entityType = formModelType(model as FormModel);
-      const associatedModels = entityType == null ? {} : await this.buildFormModels(entityType, model);
-      await this.updateModelFromForm(model, form, answers, associatedModels);
+      await this.updateModelFromForm(model, form, answers);
     }
   }
 
@@ -100,8 +96,7 @@ export class FormDataService {
       .findOne({ attributes: ["uuid", "content", "feedback", "feedbackFields"] });
     const hasURFeedback = currentUpdateRequest?.feedback != null || currentUpdateRequest?.feedbackFields != null;
     const { feedback, feedbackFields } = hasURFeedback ? currentUpdateRequest : entity;
-    const answers =
-      currentUpdateRequest?.content ?? (await this.getAnswers(form, await this.buildFormModels(entityType, entity)));
+    const answers = currentUpdateRequest?.content ?? (await this.getAnswers(form, { [entityType]: entity }));
     return populateDto(new FormDataDto(), {
       entityType,
       entityUuid: entity.uuid,
@@ -247,20 +242,6 @@ export class FormDataService {
     }
 
     return document;
-  }
-
-  async buildFormModels(modelType: FormModelType, entity: EntityModel): Promise<FormModels> {
-    const models: FormModels = { [modelType]: entity };
-
-    if (modelType === "projectReports") {
-      const report = entity as ProjectReport;
-      report.project ??= (await report.$get("project")) ?? null;
-      if (report.project != null) {
-        models.projects = report.project;
-      }
-    }
-
-    return models;
   }
 
   async getAnswers(form: Form, models: FormModels, answersModel?: { answers: Dictionary<unknown> | null }) {
