+++ apps/entity-service/src/repl/oneOff/index.ts
@@ -36,4 +36,4 @@ export { useQuestionName } from "./2026-06-25-useQuestionName";
 export { importTerraFundProjects, importTerraFundTreeSpecies } from "./2026-07-01-importTerraFundContractData";
 export { seedAboutSections } from "./2026-07-13-seedAboutSections";
 export { updateDisturbanceReportMonetaryDamageField } from "./2026-07-15-updateDisturbanceReportMonetaryDamagefield";
-export { fundoFloraProjectReportBioeconomyQuestions } from "./2026-07-16-fundoFloraProjectReportBioeconomyQuestions";
+export { fundoFloraProjectReportBioeconomyOptionLists } from "./2026-07-16-fundoFloraProjectReportBioeconomyOptionLists";
